test_scraping.py
#!/usr/bin/env python3
"""
Test script for Telecom RAG scraping functionality
"""

import os
import sys
import logging
from dotenv import load_dotenv

# Add the project root to the path
sys.path.append(os.path.dirname(os.path.abspath(__file__)))

from scraper.crawler import PlanCrawler
from scraper.extractor import PlanExtractor
from scraper.processor import DataProcessor
from knowledge_base.embedder import PlanEmbedder
from knowledge_base.vector_store import VectorStore

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

def test_scraping():
    """Test the complete scraping pipeline"""
    print("🧪 Testing Telecom RAG Scraping Pipeline")
    print("=" * 50)
    
    # Load environment variables
    load_dotenv()
    openai_api_key = os.getenv("OPENAI_API_KEY")
    
    if not openai_api_key:
        print("❌ OpenAI API key not found. Please set OPENAI_API_KEY in .env file")
        return False
    
    try:
        # Test 1: Web Crawler
        print("\n📡 Test 1: Web Crawler")
        crawler = PlanCrawler()
        
        # Test Verizon scraping
        print("   Testing Verizon scraping...")
        html_content = crawler.crawl_provider("Verizon")
        
        if html_content:
            print(f"   ✅ Successfully crawled {len(html_content)} characters")
            print(f"   📊 Content preview: {html_content[:200]}...")
        else:
            print("   ⚠️ No content retrieved, but this is expected in some environments")
        
        # Test 2: Plan Extractor
        print("\n🔍 Test 2: Plan Extractor")
        extractor = PlanExtractor()
        
        if html_content:
            plans = extractor.extract_plans(html_content, "Verizon")
            print(f"   ✅ Extracted {len(plans)} plans")
            
            if plans:
                print("   📱 Sample plan:")
                sample_plan = plans[0]
                print(f"      Name: {sample_plan.name}")
                print(f"      Price: {sample_plan.price}")
                print(f"      Data: {sample_plan.data}")
                print(f"      Features: {len(sample_plan.features)} features")
        else:
            print("   ⏭️ Skipping extraction test (no HTML content)")
        
        # Test 3: Data Processor
        print("\n⚙️ Test 3: Data Processor")
        processor = DataProcessor()
        
        # Test with sample data if no scraped data
        if not html_content or not 'plans' in locals() or not plans:
            print("   Using sample data for processor test...")
            from scraper.extractor import PlanData
            sample_plans = [
                PlanData(
                    name="Test Plan",
                    price="$75/month",
                    data="Unlimited",
                    features=["5G access", "HD streaming"],
                    url="https://test.com",
                    provider="Verizon"
                )
            ]
            processed_plans = processor.process_plans(sample_plans)
        else:
            processed_plans = processor.process_plans(plans)
        
        print(f"   ✅ Processed {len(processed_plans)} plans")
        if processed_plans:
            print("   📊 Sample processed plan:")
            sample = processed_plans[0]
            print(f"      Name: {sample['name']}")
            print(f"      Price: {sample['price']}")
            print(f"      Price Numeric: ${sample['price_numeric']:.2f}")
            print(f"      Data: {sample['data']}")
        
        # Test 4: Embeddings
        print("\n🧠 Test 4: Embeddings")
        embedder = PlanEmbedder(api_key=openai_api_key)
        
        if processed_plans:
            documents = embedder.prepare_plan_documents(processed_plans)
            print(f"   ✅ Created {len(documents)} documents")
            
            # Test embedding generation
            sample_text = "unlimited data plan with 5G"
            embedding = embedder.embed_query(sample_text)
            print(f"   ✅ Generated embedding vector of length {len(embedding)}")
        
        # Test 5: Vector Store
        print("\n📋 Test 5: Vector Store")
        vector_store = VectorStore(
            embedding_function=embedder,
            persist_directory="./test_chroma_db",
            collection_name="test_plans"
        )
        
        if processed_plans:
            # Add documents
            vector_store.add_documents(documents)
            
            # Test search
            search_results = vector_store.similarity_search("unlimited data plan", k=2)
            print(f"   ✅ Search returned {len(search_results)} results")
            
            # Get collection stats
            stats = vector_store.get_collection_stats()
            print(f"   📊 Collection stats: {stats}")
        
        # Test 6: Complete Pipeline
        print("\n🚀 Test 6: Complete Pipeline")
        
        # Test the full scrape and process pipeline
        complete_plans = processor.get_provider_data("Verizon", force_refresh=False)
        print(f"   ✅ Complete pipeline returned {len(complete_plans)} plans")
        
        print("\n🎉 All tests completed successfully!")
        print("\n📈 Summary:")
        print(f"   - Web crawling: {'Working' if html_content else 'Limited (environment dependent)'}")
        print(f"   - Plan extraction: {'Working' if 'plans' in locals() and plans else 'Using fallback'}")
        print(f"   - Data processing: Working ({len(processed_plans) if processed_plans else 0} plans)")
        print(f"   - Embeddings: Working")
        print(f"   - Vector store: Working")
        print(f"   - Complete pipeline: Working")
        
        return True
        
    except Exception as e:
        print(f"\n❌ Test failed with error: {str(e)}")
        logger.error(f"Test error: {str(e)}", exc_info=True)
        return False
    
    finally:
        # Cleanup test database
        try:
            import shutil
            if os.path.exists("./test_chroma_db"):
                shutil.rmtree("./test_chroma_db")
                print("\n🧽 Cleaned up test database")
        except Exception as e:
            print(f"\n⚠️ Cleanup warning: {str(e)}")

def test_individual_components():
    """Test individual components in isolation"""
    print("\n🔧 Testing Individual Components")
    print("=" * 40)
    
    # Test crawler initialization
    try:
        crawler = PlanCrawler()
        print("✅ Crawler initialization: Success")
    except Exception as e:
        print(f"❌ Crawler initialization: Failed - {str(e)}")
    
    # Test extractor initialization
    try:
        extractor = PlanExtractor()
        print("✅ Extractor initialization: Success")
    except Exception as e:
        print(f"❌ Extractor initialization: Failed - {str(e)}")
    
    # Test processor initialization
    try:
        processor = DataProcessor()
        print("✅ Processor initialization: Success")
    except Exception as e:
        print(f"❌ Processor initialization: Failed - {str(e)}")
    
    # Test embedder initialization
    try:
        load_dotenv()
        openai_api_key = os.getenv("OPENAI_API_KEY")
        if openai_api_key:
            embedder = PlanEmbedder(api_key=openai_api_key)
            print("✅ Embedder initialization: Success")
        else:
            print("⚠️ Embedder initialization: Skipped (no API key)")
    except Exception as e:
        print(f"❌ Embedder initialization: Failed - {str(e)}")

def main():
    """Main test function"""
    print("🚀 Starting Telecom RAG Tests")
    print("=" * 60)
    
    # Check prerequisites
    print("🔍 Checking Prerequisites")
    
    # Check Python version
    import sys
    python_version = sys.version_info
    if python_version >= (3, 8):
        print(f"✅ Python version: {python_version.major}.{python_version.minor}.{python_version.micro}")
    else:
        print(f"❌ Python version too old: {python_version.major}.{python_version.minor}.{python_version.micro} (need 3.8+)")
        return False
    
    # Check required packages
    required_packages = ['streamlit', 'openai', 'chromadb', 'playwright', 'beautifulsoup4']
    missing_packages = []
    
    for package in required_packages:
        try:
            __import__(package)
            print(f"✅ Package {package}: Installed")
        except ImportError:
            print(f"❌ Package {package}: Missing")
            missing_packages.append(package)
    
    if missing_packages:
        print(f"\n⚠️ Missing packages: {', '.join(missing_packages)}")
        print("Run: pip install -r requirements.txt")
        return False
    
    # Run component tests
    test_individual_components()
    
    # Run full pipeline test
    success = test_scraping()
    
    if success:
        print("\n🎆 All tests passed! The system is ready to use.")
        print("\n🚀 To run the application:")
        print("   streamlit run app.py")
    else:
        print("\n⚠️ Some tests failed. Check the logs above for details.")
    
    return success

if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)
