requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
chromadb>=0.4.18
pydantic==2.5.2
streamlit==1.28.2
python-dotenv==1.0.0
groq==0.4.0
lxml==4.9.3
html5lib==1.1
numpy==1.24.3
pandas==2.0.3
aiohttp==3.8.5
httpx==0.24.1


